https://public.tableau.com/views/SeoulBikeRentals_16869331659740/Sheet1?:language=en-US&:display_count=n&:origin=viz_share_link
Data Visualization Project in Tableau

Project Scenario: I am a data professional consulting for the transportation department in Seoul, South Korea. Client's contact is the director of the department. For the next project, the director requests to analyze the rented bicycle data they've gathered in 2018 in order to learn the best time of day to have their maintenance teams repair bikes. The director wants the maintenance work to occur during lower traffic times, when bicycle rental numbers are minimal, but still during standard working hours (8 a.m. to 5 p.m). My task is to use Tableau to design a data visualization that illustrates the low-traffic times of standard workdays in order to remove bicycles from circulation for maintenance.